python/requirements.txt
astroid==2.14.2
atlassian-python-api==3.32.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.2.1
Deprecated==1.2.13
dill==0.3.6
exceptiongroup==1.1.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pep8==1.7.1
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.10.0
pylint==2.16.2
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.8
zipp==3.11.0
